# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/register": {
      "post": {
        "description": "This route is used by users or authors to register a new account on the platform. It takes a first name, last name, email, password, and role. Upon success, it creates the user record.",
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created successfully"
          },
          "409": {
            "description": "Email already registered"
          }
        },
        "summary": "Create a new account",
        "tags": [
          "Users"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "description": "Authenticates a user using email and password. Returns an access token and refresh token upon success. Used for standard login flow.",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful"
          }
        },
        "summary": "Login with email and password",
        "tags": [
          "Users"
        ]
      }
    },
    "/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP sent if email exists"
          }
        },
        "summary": "Request password reset OTP",
        "tags": [
          "Users"
        ]
      }
    },
    "/auth/resend-password-reset": {
      "post": {
        "operationId": "AuthController_resendPasswordReset",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP sent if email exists"
          }
        },
        "summary": "Resend password reset OTP",
        "tags": [
          "Users"
        ]
      }
    },
    "/auth/verify-password-reset-otp": {
      "post": {
        "operationId": "AuthController_verifyPasswordResetOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyPasswordResetOtpRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP verified successfully"
          }
        },
        "summary": "Verify a password reset OTP",
        "tags": [
          "Users"
        ]
      }
    },
    "/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset successfully"
          }
        },
        "summary": "Reset password with OTP",
        "tags": [
          "Users"
        ]
      }
    },
    "/auth/change-password": {
      "post": {
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed successfully"
          }
        },
        "summary": "Change password (authenticated)",
        "tags": [
          "Users"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tokens refreshed successfully"
          }
        },
        "summary": "Refresh access token",
        "tags": [
          "Users"
        ]
      }
    },
    "/auth/verify-email": {
      "post": {
        "operationId": "AuthController_verifyEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email verified successfully"
          }
        },
        "summary": "Verify email with code",
        "tags": [
          "Users"
        ]
      }
    },
    "/auth/resend-verification": {
      "post": {
        "operationId": "AuthController_resendVerification",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResendVerificationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Code sent if verification is required"
          }
        },
        "summary": "Resend email verification code",
        "tags": [
          "Users"
        ]
      }
    },
    "/auth/google": {
      "post": {
        "operationId": "AuthController_googleLogin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleTokenRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Google authentication successful"
          }
        },
        "summary": "Login or register with Google ID token",
        "tags": [
          "Users"
        ]
      },
      "get": {
        "operationId": "AuthController_googleRedirect",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Redirect to Google OAuth consent screen",
        "tags": [
          "Users"
        ]
      }
    },
    "/auth/google/callback": {
      "get": {
        "operationId": "AuthController_googleCallback",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Google authentication successful"
          }
        },
        "summary": "Google OAuth callback",
        "tags": [
          "Users"
        ]
      }
    },
...
    "/admin/users/authors": {
      "get": {
        "operationId": "AdminUsersController_getAuthors",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ACTIVE",
                "INACTIVE",
                "SUSPENDED"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated author list"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get all authors with filtering and pagination",
        "tags": [
          "Admin Users"
        ]
      }
    },
    "/admin/users/authors/{id}": {
      "get": {
        "operationId": "AdminUsersController_getAuthor",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Author details"
          },
          "404": {
            "description": "Author not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get a single author",
        "tags": [
          "Admin Users"
        ]
      }
    },
    "/admin/users/{id}/approve": {
      "patch": {
        "operationId": "AdminUsersController_approveUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User approved successfully"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Approve a user (set status to ACTIVE)",
        "tags": [
          "Admin Users"
        ]
      }
    },
    "/admin/users/{id}/suspend": {
      "patch": {
        "operationId": "AdminUsersController_suspendUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User suspended successfully"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Suspend a user (set status to SUSPENDED)",
        "tags": [
          "Admin Users"
        ]
      }
    },
...
  },
  "info": {
    "title": "NestJS API",
    "description": "Production-ready NestJS API with Prisma, PostgreSQL, Authentication, Logging, Monitoring, and more.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT access token",
        "in": "header"
      }
    },
    "schemas": {
...
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "email": {
            "type": "string",
            "example": "john@example.com"
          },
          "password": {
            "type": "string",
            "example": "StrongPass123!"
          },
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "ADMIN",
              "MODERATOR",
              "SUPERADMIN",
              "READER",
              "AUTHOR"
            ],
            "example": "READER"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "email",
          "password",
          "role"
        ]
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "john@example.com"
          },
          "password": {
            "type": "string",
            "example": "StrongPass123!"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "ForgotPasswordRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "john@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "VerifyPasswordResetOtpRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "john@example.com"
          },
          "otp": {
            "type": "string",
            "example": "123456"
          }
        },
        "required": [
          "email",
          "otp"
        ]
      },
      "ResetPasswordRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "john@example.com"
          },
          "otp": {
            "type": "string",
            "example": "123456"
          },
          "password": {
            "type": "string",
            "example": "NewStrongPass123!"
          }
        },
        "required": [
          "email",
          "otp",
          "password"
        ]
      },
      "ChangePasswordRequest": {
        "type": "object",
        "properties": {
          "oldPassword": {
            "type": "string",
            "example": "OldPass123!"
          },
          "newPassword": {
            "type": "string",
            "example": "NewPass123!"
          }
        },
        "required": [
          "oldPassword",
          "newPassword"
        ]
      },
      "RefreshTokenRequest": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "VerifyEmailRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "john@example.com"
          },
          "code": {
            "type": "string",
            "example": "123456"
          }
        },
        "required": [
          "email",
          "code"
        ]
      },
      "ResendVerificationRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "john@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "GoogleTokenRequest": {
        "type": "object",
        "properties": {
          "idToken": {
            "type": "string",
            "description": "Google ID token from Google Sign-In"
          }
        },
        "required": [
          "idToken"
        ]
      },
...
    }
  }
}